+++ FAManagementStudio/ViewModels/Commons/ViewModelBase.cs
@@ -1,13 +1,18 @@
 ﻿using FAManagementStudio.Common;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace FAManagementStudio.ViewModels
 {
     public class ViewModelBase : BindableBase
     {
         protected Messenger MessengerInstance { get; set; } = Messenger.Instance;
+    }
+
+    public interface ITableViewModel
+    {
+        string TableName { get; }
+        string GetDdl(DbViewModel dbVm);
+        List<ColumViewMoodel> Colums { get; }
+        TableKind Kind { get; }
     }
 }
